getall user without current user

+++ src/services/user.ts
@@ -1,7 +1,6 @@
 import { col, fn, Op, where as sequelizeWhere } from 'sequelize';
 import { User } from '../models/usersModel';
 import { RegisterInput } from '../schemas/usersSchema.js';
-import { number } from 'zod';
 export interface InputUser {
   firstName: string;
   lastName: string;
@@ -10,6 +9,9 @@ export interface InputUser {
 }
 const createUser = async (userData: RegisterInput) => {
   try {
+    if (!userData) {
+      throw new Error('User data missing');
+    }
     const user = await User.create({
       firstName: userData.firstName,
       lastName: userData.lastName,
@@ -21,15 +23,17 @@ const createUser = async (userData: RegisterInput) => {
     });
     const { password, ...returnedUser } = user.dataValues;
     return returnedUser;
-
-  } catch(err){
-    console.log(err)
-    throw new Error('failed to create user')
+  } catch (err) {
+    console.error('failed to create user', err);
+    throw err;
   }
 };
 
 const createInactiveUser = async (userData: InputUser) => {
   try {
+    if (!userData) {
+      throw new Error('User data missing');
+    }
     const user = await User.create({
       firstName: userData.firstName,
       lastName: userData.lastName,
@@ -53,12 +57,12 @@ const getUserById = async (id: number) => {
 
     return user;
   } catch (err) {
-    console.log(err);
-    throw new Error('Error fetching data');
+    console.error('Error fetching data', err);
+    throw err;
   }
 };
 
-const getAllUsers = async (options: { search?: string; page?: number; offset?: number }) => {
+const getAllUsers = async (options: { search?: string; page?: number; offset?: number }, currentUserId: number) => {
   const { search, page = 1, offset = 5 } = options;
   const limit = offset;
   const offsetNum = (page - 1) * limit;
@@ -68,8 +72,7 @@ const getAllUsers = async (options: { search?: string; page?: number; offset?: n
   let where = {};
   if (search) {
     where = sequelizeWhere(fn('concat', col('firstName'), ' ', col('lastName')), {
-      [Op.iLike]: `%${search}%`,
-    });
+      [Op.iLike]: `%${search}%,`});
   }
 
   const total = await User.count({ where });
@@ -81,59 +84,71 @@ const getAllUsers = async (options: { search?: string; page?: number; offset?: n
     offset: offsetNum,
     order: [['id', 'DESC']],
   });
-  return await { users: result, total, page, pageSize: limit, totalPages: Math.ceil(total / limit) };
+  const data = result.filter(u => u.dataValues.id != currentUserId)
+  
+  return await { users: data, total, page, pageSize: limit, totalPages: Math.ceil(total / limit) };
 };
 
 const updateUser = async (
   id: number,
   updatedData: Partial<Pick<User, 'firstName' | 'lastName' | 'phoneNumber' | 'password'>>,
 ) => {
-  const user = await User.findByPk(id);
-  if (!user) return null;
-  if (updatedData.firstName !== undefined) user.firstName = updatedData.firstName;
-  if (updatedData.lastName !== undefined) user.lastName = updatedData.lastName;
-  if (updatedData.phoneNumber !== undefined) user.phoneNumber = updatedData.phoneNumber;
-  if (updatedData.password !== undefined) user.password = updatedData.password;
-  await user.save();
+  try {
+    const user = await User.findByPk(id);
+    if (!user) return null;
+    if (updatedData.firstName !== undefined) user.firstName = updatedData.firstName;
+    if (updatedData.lastName !== undefined) user.lastName = updatedData.lastName;
+    if (updatedData.phoneNumber !== undefined) user.phoneNumber = updatedData.phoneNumber;
+    if (updatedData.password !== undefined) user.password = updatedData.password;
+    await user.save();
 
-  return user.dataValues;
+    return user.dataValues;
+  } catch (err) {
+    console.error('Fail to update user', err);
+    throw err;
+  }
 };
 
 const getUserByEmail = async (email: string, includeDeleted: boolean = false) => {
   try {
     const user = await User.findOne({
       where: { email },
-      paranoid: !includeDeleted
+      paranoid: !includeDeleted,
     });
 
     if (!user) return null;
     return user.dataValues;
   } catch (err) {
     console.error('Error fetching user by email:', err);
-    throw new Error('Error fetching user data');
+    throw err;
   }
 };
 
-
 const softDeleteUser = async (userId: number) => {
   try {
-    const user = await User.findByPk(userId)
-    if(!user) {
-      throw Error('User not found')
+    if (!userId) {
+      throw new Error('User id missing');
+    }
+    const user = await User.findByPk(userId);
+    if (!user) {
+      throw Error('User not found');
     }
     user.isActive = false;
     user.password = null;
-    await user.save()
+    await user.save();
     const deletedRows = await User.destroy({ where: { id: userId } });
     return deletedRows > 0;
-  } catch (error) {
-    console.error('Error soft deleting user:', error);
-    throw error;
+  } catch (err) {
+    console.error('Error soft deleting user:', err);
+    throw err;
   }
 };
 
 const restoreUser = async (userId: number) => {
   try {
+    if (!userId) {
+      throw new Error('User id missing');
+    }
     const user = await User.findByPk(userId, { paranoid: false });
     if (!user) {
       throw new Error('User not found');
@@ -144,24 +159,32 @@ const restoreUser = async (userId: number) => {
     }
 
     await user.restore();
-  
+
     return true;
-  } catch (error) {
-    console.error('Error restoring user:', error);
-    throw error;
+  } catch (err) {
+    console.error('Error restoring user:', err);
+    throw err;
   }
 };
 
 const updateUserPasswordActiveStatus = async (updatedUser: User) => {
-  const user = await User.findByPk(updatedUser.id);
-  if (!user) return null;
-  
-  if (updatedUser.password !== undefined) user.password = updatedUser.password;
-  if (updatedUser.isActive !== undefined) user.isActive = updatedUser.isActive;
-  
-  await user.save();
-  
-  return user.dataValues;
+  try {
+    if (!updateUser) {
+      throw new Error('Updated user data missing ');
+    }
+    const user = await User.findByPk(updatedUser.id);
+    if (!user) return null;
+
+    if (updatedUser.password !== undefined) user.password = updatedUser.password;
+    if (updatedUser.isActive !== undefined) user.isActive = updatedUser.isActive;
+
+    await user.save();
+
+    return user.dataValues;
+  } catch (err) {
+    console.error('Failed to update user');
+    throw err;
+  }
 };
 
 export default {
@@ -173,5 +196,5 @@ export default {
   updateUserPasswordActiveStatus,
   createInactiveUser,
   softDeleteUser,
-  restoreUser
+  restoreUser,
 };
